+++ R/AcqOptimizerDirect.R
@@ -67,7 +67,6 @@ AcqOptimizerDirect = R6Class(
         xtol_abs = p_dbl(default = 0, lower = 0, upper = Inf, special_vals = list(-1)),
         ftol_rel = p_dbl(default = 0, lower = 0, upper = Inf, special_vals = list(-1)),
         ftol_abs = p_dbl(default = 0, lower = 0, upper = Inf, special_vals = list(-1)),
-        minf_max = p_dbl(default = -Inf),
         catch_errors = p_lgl(init = TRUE)
       )
       private$.param_set = param_set